backend/requirements.txt
fastapi
uvicorn
transformers
torch
shap
google-generativeai
python-dotenv
pydantic
scipy
pandas
Requests
matplotlib
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
psycopg2-binary
motor
certifi
imap-tools
